requirements.txt
Django==5.2.6
django-environ==0.12.0
mysqlclient==2.2.7
pillow==11.3.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
